moved interfaces to a separate interfaces folder
