+++ angular/src/interfaces/amenity.ts
