+++ angular/src/interfaces/room.ts
