+++ angular/src/interfaces/trainingcenter.ts
